1. The code organized. 2. Git ignore file added.
