+++ Net_Tutorial/SwiftyJsonAndUrlConnection(Swift3)/SwiftyJsonDemo/Constants.swift
@@ -1,0 +1,18 @@
+//
+//  Constants.swift
+//  SwiftyJsonDemo
+//
+//  Created by Rocoo on 2016/11/9.
+//  Copyright © 2016年 Rocoo. All rights reserved.
+//
+
+import Foundation
+
+let timeOut: TimeInterval = 10
+
+let connectionSuccess = 200
+
+let errorDomain = "TaoG"
+
+let errorNetwork = -1009
+let errorTimeout = -1001
