+++ Net_Tutorial/SwiftyJsonAndUrlConnection(Swift3)/SwiftyJsonDemo/Extension.swift
@@ -1,0 +1,19 @@
+//
+//  Extension.swift
+//  SwiftyJsonDemo
+//
+//  Created by Rocoo on 2016/11/9.
+//  Copyright © 2016年 Rocoo. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+extension UIViewController{
+    func connectionFailAlert(){
+        let alert = UIAlertController(title: "Connection Fail.", message: "Please check your internet and try it again.", preferredStyle: .alert)
+        alert.addAction(UIAlertAction(title: "ok", style: .default, handler: nil))
+        
+        self.present(alert, animated: true, completion: nil)
+    }
+}
